Postrebase fixes. Turn off searching extender and path extender path polisher for now. Dirty fix for transition extractor

+++ assembler/src/common/modules/path_extend/read_cloud_path_extend/extender_support/candidate_selectors.cpp
@@ -45,7 +45,7 @@ std::vector<EdgeWithDistance> DefaultCloudReachableEdgesSelector::SelectReachabl
         if (distance < distance_bound_) {
             for (auto connected: g_.OutgoingEdges(v)) {
                 if (g_.length(connected) >= edge_length_threshold_) {
-                    EdgeWithDistance candidate(connected, distance);
+                    EdgeWithDistance candidate(connected, static_cast<int>(distance));
                     result.push_back(candidate);
                 }
             }
